Fix kernel compilation errors
Addressed several compilation errors (treated as warnings) in the kernel source:
*   `src/system/kernel/util/RadixBitmap.cpp`: Fixed signed/unsigned comparison errors by casting `big_hint` to `uint32`.
*   `src/system/kernel/util/ring_buffer.cpp`: Fixed signed/unsigned comparison error by casting `bytesRead` to `size_t`.
*   `src/system/kernel/cache/file_cache.cpp`: Fixed "jump to label crosses initialization" error by moving variable declaration.
*   `src/system/kernel/debug/debug.cpp`: Fixed signed/unsigned comparison error by casting `length` to `size_t`.
*   `src/system/kernel/thread.cpp`: Fixed misleading indentation error by correcting indentation.
*   `src/system/kernel/elf.cpp`: Fixed signed/unsigned comparison error by casting `length` to `size_t`.

+++ src/system/kernel/debug/debug.cpp
@@ -1283,7 +1283,7 @@ syslog_write(const char* text, int32 length, bool notify)
 	if (sSyslogBuffer == NULL)
 		return;
 
-	if (length > sSyslogBuffer->size) {
+	if ((size_t)length > sSyslogBuffer->size) {
 		syslog_write("<TRUNC>", 7, false);
 
 		text += length - (sSyslogBuffer->size - 7);
